pyproject.toml
[project]
name = "holly-grace"
version = "1.0.0"
description = "Holly Grace — autonomous e-commerce agent orchestration platform"
requires-python = ">=3.11"
dependencies = [
    # LangChain ecosystem
    "langgraph>=0.2.0,<1.0.0",
    "langchain>=0.3.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "langchain-anthropic>=0.3.0,<1.0.0",
    "langchain-ollama>=0.2.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langchain-chroma>=0.2.0,<1.0.0",
    "chromadb-client>=0.5.0,<1.0.0",
    "langsmith>=0.2.0,<1.0.0",
    "langserve[all]>=0.3.0,<1.0.0",
    # Graph checkpointing
    "langgraph-checkpoint-postgres>=2.0.0,<3.0.0",
    "psycopg[binary]>=3.1.0,<4.0.0",
    # External service SDKs
    "stripe-agent-toolkit>=0.1.0,<1.0.0",
    "stripe>=8.0.0,<12.0.0",
    "httpx>=0.27.0,<1.0.0",
    # Vector DB (client-only; server runs in Docker)
    # Scheduling
    "apscheduler>=3.10.0,<4.0.0",
    # Infrastructure
    "redis>=5.0.0,<6.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "fastapi>=0.109.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "sse-starlette>=1.6.0,<3.0.0",
    # APS (Adaptive Partition Selection)
    "numpy>=1.24.0,<3.0.0",
    # Security
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "slowapi>=0.1.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "ruff>=0.4.0,<1.0.0",
    "mypy>=1.7.0,<2.0.0",
    "hypothesis>=6.100.0,<7.0.0",
    "freezegun>=1.2.0,<2.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
